requirements.txt
crewai
crewai-tools
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
pandas
python-dotenv
google-auth
openai-whisper
assemblyai
google-generativeai
email-validator